chore: extract methods for agent checks to its own file
