+++ e2e/_suites/fleet/fleet.go
@@ -11,7 +11,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/Jeffail/gabs/v2"
@@ -22,11 +21,9 @@ import (
 	"github.com/elastic/e2e-testing/internal/common"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/elasticsearch"
-	"github.com/elastic/e2e-testing/internal/installer"
 	"github.com/elastic/e2e-testing/internal/kibana"
 	"github.com/elastic/e2e-testing/internal/utils"
 
-	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -68,47 +65,6 @@ func (fts *FleetTestSuite) getDeployer() deploy.Deployment {
 		return fts.dockerDeployer
 	}
 	return fts.deployer
-}
-
-func (fts *FleetTestSuite) tagsAreInTheElasticAgentIndex() error {
-	var tagsArray []string
-	//ex of flags  "--tag production,linux" or "--tag=production,linux"
-	if fts.ElasticAgentFlags != "" {
-		tags := strings.TrimPrefix(fts.ElasticAgentFlags, "--tag")
-		tags = strings.TrimPrefix(tags, "=")
-		tags = strings.ReplaceAll(tags, " ", "")
-		tagsArray = strings.Split(tags, ",")
-	}
-	if len(tagsArray) == 0 {
-		return errors.Errorf("no tags were found, ElasticAgentFlags value %s", fts.ElasticAgentFlags)
-	}
-
-	var tagTerms []map[string]interface{}
-	for _, tag := range tagsArray {
-		tagTerms = append(tagTerms, map[string]interface{}{
-			"term": map[string]interface{}{
-				"tags": tag,
-			},
-		})
-	}
-
-	query := map[string]interface{}{
-		"query": map[string]interface{}{
-			"bool": map[string]interface{}{
-				"must": tagTerms,
-			},
-		},
-	}
-
-	indexName := ".fleet-agents"
-
-	_, err := elasticsearch.WaitForNumberOfHits(context.Background(), indexName, query, 1, 3*time.Minute)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err,
-		}).Warn(elasticsearch.WaitForIndices())
-	}
-	return err
 }
 
 // bootstrapFleet this method creates the runtime dependencies for the Fleet test suite, being of special
@@ -260,122 +216,6 @@ func (fts *FleetTestSuite) setup() error {
 	return nil
 }
 
-func (fts *FleetTestSuite) theAgentIsListedInFleetWithStatus(desiredStatus string) error {
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
-	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
-	err := theAgentIsListedInFleetWithStatus(fts.currentContext, desiredStatus, manifest.Hostname)
-	if err != nil {
-		return err
-	}
-	if desiredStatus == "online" {
-		//get Agent Default Key
-		err := fts.theAgentGetDefaultAPIKey()
-		if err != nil {
-			return err
-		}
-	}
-	return err
-}
-
-func theAgentIsListedInFleetWithStatus(ctx context.Context, desiredStatus string, hostname string) error {
-	log.Tracef("Checking if agent is listed in Fleet as %s", desiredStatus)
-
-	kibanaClient, err := kibana.NewClient()
-	if err != nil {
-		return err
-	}
-	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute * 2
-	retryCount := 1
-
-	exp := utils.GetExponentialBackOff(maxTimeout)
-
-	agentOnlineFn := func() error {
-		agentID, err := kibanaClient.GetAgentIDByHostname(ctx, hostname)
-		if err != nil {
-			retryCount++
-			return err
-		}
-
-		if agentID == "" {
-			// the agent is not listed in Fleet
-			if desiredStatus == "offline" || desiredStatus == "inactive" {
-				log.WithFields(log.Fields{
-					"elapsedTime": exp.GetElapsedTime(),
-					"hostname":    hostname,
-					"retries":     retryCount,
-					"status":      desiredStatus,
-				}).Info("The Agent is not present in Fleet, as expected")
-				return nil
-			}
-
-			retryCount++
-			return fmt.Errorf("the agent is not present in Fleet in the '%s' status, but it should", desiredStatus)
-		}
-
-		agentStatus, err := kibanaClient.GetAgentStatusByHostname(ctx, hostname)
-		isAgentInStatus := strings.EqualFold(agentStatus, desiredStatus)
-		if err != nil || !isAgentInStatus {
-			if err == nil {
-				err = fmt.Errorf("the Agent is not in the %s status yet", desiredStatus)
-			}
-
-			log.WithFields(log.Fields{
-				"agentID":         agentID,
-				"isAgentInStatus": isAgentInStatus,
-				"elapsedTime":     exp.GetElapsedTime(),
-				"hostname":        hostname,
-				"retry":           retryCount,
-				"status":          desiredStatus,
-			}).Warn(err.Error())
-
-			retryCount++
-
-			return err
-		}
-		log.WithFields(log.Fields{
-			"isAgentInStatus": isAgentInStatus,
-			"elapsedTime":     exp.GetElapsedTime(),
-			"hostname":        hostname,
-			"retries":         retryCount,
-			"status":          desiredStatus,
-		}).Info("The Agent is in the desired status")
-		return nil
-	}
-
-	err = backoff.Retry(agentOnlineFn, exp)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (fts *FleetTestSuite) theFileSystemAgentFolderIsEmpty() error {
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
-	agentInstaller, _ := installer.Attach(fts.currentContext, fts.getDeployer(), agentService, fts.InstallerType)
-
-	pkgManifest, _ := agentInstaller.Inspect()
-	cmd := []string{
-		"ls", "-l", pkgManifest.WorkDir,
-	}
-
-	content, err := agentInstaller.Exec(fts.currentContext, cmd)
-	if err != nil {
-		if content == "" || strings.Contains(content, "No such file or directory") {
-			return nil
-		}
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"installer":  agentInstaller,
-		"workingDir": pkgManifest.WorkDir,
-		"content":    content,
-	}).Debug("Agent working dir content")
-
-	return fmt.Errorf("the file system directory is not empty")
-}
-
 func (fts *FleetTestSuite) theHostIsRestarted() error {
 	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
 	err := fts.getDeployer().Stop(fts.currentContext, agentService)
@@ -392,161 +232,6 @@ func (fts *FleetTestSuite) theHostIsRestarted() error {
 
 	log.Debug("The elastic-agent service has been restarted")
 	return nil
-}
-
-func (fts *FleetTestSuite) systemPackageDashboardsAreListedInFleet() error {
-	log.Trace("Checking system Package dashboards in Fleet")
-
-	dataStreamsCount := 0
-	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
-	retryCount := 1
-
-	exp := utils.GetExponentialBackOff(maxTimeout)
-
-	countDataStreamsFn := func() error {
-		dataStreams, err := fts.kibanaClient.GetDataStreams(fts.currentContext)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"retry":       retryCount,
-				"elapsedTime": exp.GetElapsedTime(),
-			}).Warn(err.Error())
-
-			retryCount++
-
-			return err
-		}
-
-		count := len(dataStreams.Children())
-		if count == 0 {
-			err = fmt.Errorf("there are no datastreams yet")
-
-			log.WithFields(log.Fields{
-				"retry":       retryCount,
-				"dataStreams": count,
-				"elapsedTime": exp.GetElapsedTime(),
-			}).Warn(err.Error())
-
-			retryCount++
-
-			return err
-		}
-
-		log.WithFields(log.Fields{
-			"elapsedTime": exp.GetElapsedTime(),
-			"datastreams": count,
-			"retries":     retryCount,
-		}).Info("Datastreams are present")
-		dataStreamsCount = count
-		return nil
-	}
-
-	err := backoff.Retry(countDataStreamsFn, exp)
-	if err != nil {
-		return err
-	}
-
-	if dataStreamsCount == 0 {
-		err = fmt.Errorf("there are no datastreams. We expected to have more than one")
-		log.Error(err.Error())
-		return err
-	}
-
-	return nil
-}
-
-// theVersionOfThePackageIsInstalled installs a package in a version
-func (fts *FleetTestSuite) theVersionOfThePackageIsInstalled(version string, packageName string) error {
-	log.WithFields(log.Fields{
-		"package": packageName,
-		"version": version,
-	}).Trace("Checking if package version is installed")
-
-	integration, err := fts.kibanaClient.GetIntegrationByPackageName(fts.currentContext, packageName)
-	if err != nil {
-		return err
-	}
-
-	_, err = fts.kibanaClient.InstallIntegrationAssets(fts.currentContext, integration)
-	if err != nil {
-		return err
-	}
-	fts.Integration = integration
-
-	return nil
-}
-
-func (fts *FleetTestSuite) checkDataStream() error {
-	query := map[string]interface{}{
-		"query": map[string]interface{}{
-			"bool": map[string]interface{}{
-				"filter": []interface{}{
-					map[string]interface{}{
-						"exists": map[string]interface{}{
-							"field": "linux.memory.page_stats",
-						},
-					},
-					map[string]interface{}{
-						"exists": map[string]interface{}{
-							"field": "elastic_agent",
-						},
-					},
-					map[string]interface{}{
-						"range": map[string]interface{}{
-							"@timestamp": map[string]interface{}{
-								"gte": "now-1m",
-							},
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"data_stream.type": "metrics",
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"data_stream.dataset": "linux.memory",
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"data_stream.namespace": "default",
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"event.dataset": "linux.memory",
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"agent.type": "metricbeat",
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"metricset.period": 1000,
-						},
-					},
-					map[string]interface{}{
-						"term": map[string]interface{}{
-							"service.type": "linux",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	indexName := "metrics-linux.memory-default"
-
-	_, err := elasticsearch.WaitForNumberOfHits(context.Background(), indexName, query, 1, 3*time.Minute)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error": err,
-		}).Warn(elasticsearch.WaitForIndices())
-	}
-
-	return err
 }
 
 func inputs(integration string) []kibana.Input {
